Refactored sample data into more meaningful packaging for reuseability
