+++ argonaut-api/src/test/java/gov/va/api/health/argonaut/api/samples/SampleMedications.java
@@ -1,7 +1,9 @@
-package gov.va.api.health.argonaut.api;
+package gov.va.api.health.argonaut.api.samples;
 
 import static java.util.Collections.singletonList;
 
+import gov.va.api.health.argonaut.api.CodeableConcept;
+import gov.va.api.health.argonaut.api.Medication;
 import java.util.Arrays;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Delegate;
