+++ argonaut/src/test/java/gov/va/api/health/argonaut/service/controller/patient/PatientTransformerTest.java
@@ -1,49 +1,17 @@
 package gov.va.api.health.argonaut.service.controller.patient;
 
+import gov.va.api.health.argonaut.api.*;
+import gov.va.api.health.argonaut.api.samples.SamplePatients;
+import gov.va.api.health.argonaut.service.samples.SampleCdwPatients;
+import gov.va.api.health.argonaut.service.controller.Transformers;
+import org.junit.Test;
+import java.util.*;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
-import gov.va.api.health.argonaut.api.Address;
-import gov.va.api.health.argonaut.api.CodeableConcept;
-import gov.va.api.health.argonaut.api.Coding;
-import gov.va.api.health.argonaut.api.Contact;
-import gov.va.api.health.argonaut.api.ContactPoint;
-import gov.va.api.health.argonaut.api.Extension;
-import gov.va.api.health.argonaut.api.HumanName;
-import gov.va.api.health.argonaut.api.Identifier;
-import gov.va.api.health.argonaut.api.Reference;
-import gov.va.api.health.argonaut.service.controller.Transformers;
-import gov.va.dvp.cdw.xsd.model.CdwAdministrativeGenderCodes;
-import gov.va.dvp.cdw.xsd.model.CdwBirthSexCodes;
-import gov.va.dvp.cdw.xsd.model.CdwBirthsexExtension;
-import gov.va.dvp.cdw.xsd.model.CdwContactPointSystemCodes;
-import gov.va.dvp.cdw.xsd.model.CdwContactPointUseCodes;
-import gov.va.dvp.cdw.xsd.model.CdwExtensions;
-import gov.va.dvp.cdw.xsd.model.CdwExtensions.CdwExtension;
-import gov.va.dvp.cdw.xsd.model.CdwIdentifierUseCodes;
-import gov.va.dvp.cdw.xsd.model.CdwMaritalStatusCodes;
-import gov.va.dvp.cdw.xsd.model.CdwMaritalStatusSystems;
-import gov.va.dvp.cdw.xsd.model.CdwPatient103Root.CdwPatients.CdwPatient;
-import gov.va.dvp.cdw.xsd.model.CdwPatient103Root.CdwPatients.CdwPatient.CdwAddresses;
-import gov.va.dvp.cdw.xsd.model.CdwPatient103Root.CdwPatients.CdwPatient.CdwAddresses.CdwAddress;
-import gov.va.dvp.cdw.xsd.model.CdwPatient103Root.CdwPatients.CdwPatient.CdwContacts;
-import gov.va.dvp.cdw.xsd.model.CdwPatient103Root.CdwPatients.CdwPatient.CdwContacts.CdwContact;
-import gov.va.dvp.cdw.xsd.model.CdwPatientContactRelationshipCodes;
-import gov.va.dvp.cdw.xsd.model.CdwPatientContactRelationshipSystem;
-import java.math.BigInteger;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Optional;
-import javax.xml.datatype.DatatypeFactory;
-import javax.xml.datatype.XMLGregorianCalendar;
-import lombok.SneakyThrows;
-import org.junit.Test;
-
 public class PatientTransformerTest {
-
-  private XmlSampleData cdw = new XmlSampleData();
-  private PatientSampleData expectedPatient = new PatientSampleData();
+  private SampleCdwPatients cdw = new SampleCdwPatients();
+  private SamplePatients expectedPatient = SamplePatients.get();
 
   @Test
   public void addressReturnsNullForNull() {
@@ -351,488 +319,5 @@ public class PatientTransformerTest {
 
   private PatientTransformer transformer() {
     return new PatientTransformer();
-  }
-
-  static class PatientSampleData {
-
-    private DatatypeFactory datatypeFactory;
-
-    @SneakyThrows
-    private PatientSampleData() {
-      datatypeFactory = DatatypeFactory.newInstance();
-    }
-
-    List<Address> address() {
-      List<Address> addresses = new LinkedList<>();
-      addresses.add(Address.builder().state("Missing*").build());
-      List<String> line1 = new LinkedList<>();
-      line1.add("1234 Test Road");
-      line1.add("Testland");
-      line1.add("Test POSTGRES");
-      addresses.add(
-          Address.builder()
-              .line(line1)
-              .city("Testville")
-              .state("Testlina")
-              .postalCode("12345")
-              .build());
-      List<String> line2 = new LinkedList<>();
-      line2.add("9876 Fake Lane");
-      addresses.add(
-          Address.builder()
-              .line(line2)
-              .city("Fooville")
-              .state("Foolina")
-              .postalCode("98765")
-              .build());
-      return addresses;
-    }
-
-    gov.va.api.health.argonaut.api.Patient alivePatient() {
-      return gov.va.api.health.argonaut.api.Patient.builder()
-          .resourceType("Patient")
-          .id("123456789")
-          .extension(argoCdwExtensions())
-          .identifier(identifier())
-          .name(name())
-          .telecom(telecom())
-          .gender(gov.va.api.health.argonaut.api.Patient.Gender.male)
-          .birthDate("2018-11-06")
-          .deceasedBoolean(false)
-          .address(address())
-          .maritalStatus(maritalStatus())
-          .contact(contact())
-          .build();
-    }
-
-    List<Extension> argoCdwExtensions() {
-      List<Extension> CdwExtensions = new ArrayList<>(3);
-
-      List<Extension> raceCdwExtensions = new LinkedList<>();
-      raceCdwExtensions.add(
-          Extension.builder()
-              .url("ombTest")
-              .valueCoding(
-                  Coding.builder()
-                      .system("http://test-race")
-                      .code("R4C3")
-                      .display("tester")
-                      .build())
-              .build());
-      raceCdwExtensions.add(Extension.builder().url("text").valueString("tester").build());
-
-      List<Extension> ethnicityCdwExtensions = new LinkedList<>();
-      ethnicityCdwExtensions.add(
-          Extension.builder()
-              .url("ombTest")
-              .valueCoding(
-                  Coding.builder()
-                      .system("http://test-ethnicity")
-                      .code("3THN1C1TY")
-                      .display("testa")
-                      .build())
-              .build());
-      ethnicityCdwExtensions.add(Extension.builder().url("text").valueString("testa").build());
-
-      CdwExtensions.add(
-          Extension.builder().url("http://test-race").extension(raceCdwExtensions).build());
-      CdwExtensions.add(
-          Extension.builder()
-              .url("http://test-ethnicity")
-              .extension(ethnicityCdwExtensions)
-              .build());
-      CdwExtensions.add(Extension.builder().url("http://test-birthsex").valueCode("M").build());
-      return CdwExtensions;
-    }
-
-    List<Contact> contact() {
-      List<Contact> contacts = new LinkedList<>();
-
-      List<String> line1 = new LinkedList<>();
-      line1.add("123 Happy Avenue");
-      line1.add("456 Smile Drive");
-      line1.add("789 Laughter Lane");
-      contacts.add(
-          Contact.builder()
-              .relationship(relationship())
-              .name(HumanName.builder().text("DUCK, DAFFY JOHN").build())
-              .telecom(
-                  Collections.singletonList(
-                      ContactPoint.builder()
-                          .system(ContactPoint.ContactPointSystem.phone)
-                          .value("9998886666")
-                          .build()))
-              .address(
-                  Address.builder()
-                      .line(line1)
-                      .city("Happyland")
-                      .state("Happylina")
-                      .postalCode("12345")
-                      .country("USA")
-                      .build())
-              .build());
-
-      List<String> line2 = new LinkedList<>();
-      line2.add("123 Sad Avenue");
-      line2.add("456 Frown Drive");
-      line2.add("789 Weeping Lane");
-      contacts.add(
-          Contact.builder()
-              .relationship(relationship())
-              .name(HumanName.builder().text("ALICE, TEST JANE").build())
-              .telecom(
-                  Collections.singletonList(
-                      ContactPoint.builder()
-                          .system(ContactPoint.ContactPointSystem.phone)
-                          .value("1112224444")
-                          .build()))
-              .address(
-                  Address.builder()
-                      .line(line2)
-                      .city("Sadland")
-                      .state("Sadlina")
-                      .postalCode("98765")
-                      .country("USA")
-                      .build())
-              .build());
-
-      return contacts;
-    }
-
-    gov.va.api.health.argonaut.api.Patient deceasedPatient() {
-      return gov.va.api.health.argonaut.api.Patient.builder()
-          .id("123456789")
-          .extension(argoCdwExtensions())
-          .identifier(identifier())
-          .name(name())
-          .telecom(telecom())
-          .gender(gov.va.api.health.argonaut.api.Patient.Gender.male)
-          .birthDate("2018-11-06")
-          .deceasedBoolean(true)
-          .deceasedDateTime("2018-11-07")
-          .address(address())
-          .maritalStatus(maritalStatus())
-          .contact(contact())
-          .build();
-    }
-
-    List<Identifier> identifier() {
-      List<Identifier> identifiers = new LinkedList<>();
-      identifiers.add(
-          Identifier.builder()
-              .use(Identifier.IdentifierUse.usual)
-              .type(
-                  CodeableConcept.builder()
-                      .coding(
-                          Collections.singletonList(
-                              Coding.builder().system("http://test-code").code("C0D3").build()))
-                      .build())
-              .system("http://test-system")
-              .value("123456789")
-              .assigner(Reference.builder().display("tester-test-index").build())
-              .build());
-
-      return identifiers;
-    }
-
-    CodeableConcept maritalStatus() {
-      return CodeableConcept.builder()
-          .coding(
-              Collections.singletonList(
-                  Coding.builder()
-                      .system("http://hl7.org/fhir/marital-status")
-                      .code("M")
-                      .display("Married")
-                      .build()))
-          .text("testMarriage")
-          .build();
-    }
-
-    List<HumanName> name() {
-      return Collections.singletonList(
-          HumanName.builder()
-              .use(HumanName.NameUse.usual)
-              .text("FOOMAN FOO")
-              .family(Collections.singletonList("FOO"))
-              .given(Collections.singletonList("FOOMAN"))
-              .build());
-    }
-
-    List<CodeableConcept> relationship() {
-      return Collections.singletonList(
-          CodeableConcept.builder()
-              .coding(
-                  Collections.singletonList(
-                      Coding.builder()
-                          .system("http://hl7.org/fhir/patient-contact-relationship")
-                          .code("emergency")
-                          .display("Emergency")
-                          .build()))
-              .text("Emergency Contact")
-              .build());
-    }
-
-    List<ContactPoint> telecom() {
-      List<ContactPoint> telecoms = new LinkedList<>();
-      telecoms.add(
-          ContactPoint.builder()
-              .system(ContactPoint.ContactPointSystem.phone)
-              .value("9998886666")
-              .use(ContactPoint.ContactPointUse.home)
-              .build());
-      telecoms.add(
-          ContactPoint.builder()
-              .system(ContactPoint.ContactPointSystem.phone)
-              .value("1112224444")
-              .use(ContactPoint.ContactPointUse.work)
-              .build());
-      return telecoms;
-    }
-  }
-
-  private static class XmlSampleData {
-
-    private DatatypeFactory datatypeFactory;
-
-    @SneakyThrows
-    private XmlSampleData() {
-      datatypeFactory = DatatypeFactory.newInstance();
-    }
-
-    CdwBirthsexExtension CdwBirthsexExtension() {
-      CdwBirthsexExtension testCdwBirthsexExtension = new CdwBirthsexExtension();
-      testCdwBirthsexExtension.setValueCode(CdwBirthSexCodes.M);
-      testCdwBirthsexExtension.setUrl("http://test-birthsex");
-      return testCdwBirthsexExtension;
-    }
-
-    CdwAddresses addresses() {
-
-      CdwAddresses addresses = new CdwAddresses();
-
-      CdwAddress address = new CdwAddress();
-      address.setState("Missing*");
-      addresses.getAddress().add(address);
-
-      address = new CdwAddress();
-      address.setStreetAddress1("1234 Test Road");
-      address.setStreetAddress2("Testland");
-      address.setStreetAddress3("Test POSTGRES");
-      address.setCity("Testville");
-      address.setState("Testlina");
-      address.setPostalCode("12345");
-      addresses.getAddress().add(address);
-
-      address = new CdwAddress();
-      address.setStreetAddress1("9876 Fake Lane");
-      address.setCity("Fooville");
-      address.setState("Foolina");
-      address.setPostalCode("98765");
-      addresses.getAddress().add(address);
-
-      return addresses;
-    }
-
-    CdwPatient alivePatient() {
-      CdwPatient patient = new CdwPatient();
-      patient.setRowNumber(new BigInteger("1"));
-      patient.setCdwId("123456789");
-      patient.getArgoRace().add(raceCdwExtensions());
-      patient.getArgoEthnicity().add(ethnicityCdwExtensions());
-      patient.setArgoBirthsex(CdwBirthsexExtension());
-      patient.getIdentifier().add(identifier());
-      patient.setName(name());
-      patient.setTelecoms(telecoms());
-      patient.setAddresses(addresses());
-      patient.setGender(CdwAdministrativeGenderCodes.MALE);
-      patient.setBirthDate(birthdate());
-      patient.setDeceasedBoolean(false);
-      patient.setMaritalStatus(maritalStatus());
-      patient.setContacts(contacts());
-      return patient;
-    }
-
-    @SneakyThrows
-    XMLGregorianCalendar birthdate() {
-      XMLGregorianCalendar birthdate = datatypeFactory.newXMLGregorianCalendar();
-      birthdate.setYear(2018);
-      birthdate.setMonth(11);
-      birthdate.setDay(06);
-      return birthdate;
-    }
-
-    CdwContacts contacts() {
-      CdwContacts contacts = new CdwContacts();
-
-      CdwContact contact1 = new CdwContact();
-
-      CdwContact.CdwRelationship relationship1 = new CdwContact.CdwRelationship();
-
-      CdwContact.CdwRelationship.CdwCoding relationshipCoding1 =
-          new CdwContact.CdwRelationship.CdwCoding();
-      relationshipCoding1.setSystem(
-          CdwPatientContactRelationshipSystem.HTTP_HL_7_ORG_FHIR_PATIENT_CONTACT_RELATIONSHIP);
-      relationshipCoding1.setCode(CdwPatientContactRelationshipCodes.EMERGENCY);
-      relationshipCoding1.setDisplay("Emergency");
-
-      relationship1.setCoding(relationshipCoding1);
-      relationship1.setText("Emergency Contact");
-
-      contact1.setRelationship(relationship1);
-      contact1.setName("DUCK, DAFFY JOHN");
-      contact1.setPhone("9998886666");
-      contact1.setStreetAddress1("123 Happy Avenue");
-      contact1.setStreetAddress2("456 Smile Drive");
-      contact1.setStreetAddress3("789 Laughter Lane");
-      contact1.setCity("Happyland");
-      contact1.setState("Happylina");
-      contact1.setPostalCode("12345");
-      contact1.setCountry("USA");
-      contacts.getContact().add(contact1);
-
-      CdwContact contact2 = new CdwContact();
-
-      CdwContact.CdwRelationship relationship2 = new CdwContact.CdwRelationship();
-
-      CdwContact.CdwRelationship.CdwCoding relationshipCoding2 =
-          new CdwContact.CdwRelationship.CdwCoding();
-      relationshipCoding2.setSystem(
-          CdwPatientContactRelationshipSystem.HTTP_HL_7_ORG_FHIR_PATIENT_CONTACT_RELATIONSHIP);
-      relationshipCoding2.setCode(CdwPatientContactRelationshipCodes.EMERGENCY);
-      relationshipCoding2.setDisplay("Emergency");
-
-      relationship2.setCoding(relationshipCoding2);
-      relationship2.setText("Emergency Contact");
-
-      contact2.setRelationship(relationship2);
-      contact2.setName("ALICE, TEST JANE");
-      contact2.setPhone("1112224444");
-      contact2.setStreetAddress1("123 Sad Avenue");
-      contact2.setStreetAddress2("456 Frown Drive");
-      contact2.setStreetAddress3("789 Weeping Lane");
-      contact2.setCity("Sadland");
-      contact2.setState("Sadlina");
-      contact2.setPostalCode("98765");
-      contact2.setCountry("USA");
-      contacts.getContact().add(contact2);
-
-      return contacts;
-    }
-
-    CdwPatient deadPatient() {
-      CdwPatient patient = new CdwPatient();
-      patient.setDeceasedBoolean(true);
-      patient.setDeceasedDateTime(deceasedDateTime());
-      return patient;
-    }
-
-    XMLGregorianCalendar deceasedDateTime() {
-      XMLGregorianCalendar deceasedDate = datatypeFactory.newXMLGregorianCalendar();
-      deceasedDate.setYear(2018);
-      deceasedDate.setMonth(11);
-      deceasedDate.setDay(07);
-      return deceasedDate;
-    }
-
-    CdwExtensions ethnicityCdwExtensions() {
-      CdwExtensions extensions = new CdwExtensions();
-      extensions.setUrl("http://test-ethnicity");
-
-      CdwExtension extension1 = new CdwExtension();
-      extension1.setUrl("ombTest");
-
-      CdwExtension.CdwValueCoding valueCoding = new CdwExtension.CdwValueCoding();
-      valueCoding.setSystem("http://test-ethnicity");
-      valueCoding.setCode("3THN1C1TY");
-      valueCoding.setDisplay("testa");
-      extension1.setValueCoding(valueCoding);
-
-      extensions.getExtension().add(extension1);
-
-      CdwExtension extension2 = new CdwExtension();
-      extension2.setUrl("text");
-      extension2.setValueString("testa");
-
-      extensions.getExtension().add(extension2);
-      return extensions;
-    }
-
-    CdwPatient.CdwIdentifier identifier() {
-      CdwPatient.CdwIdentifier identifier = new CdwPatient.CdwIdentifier();
-      identifier.setUse(CdwIdentifierUseCodes.USUAL);
-      CdwPatient.CdwIdentifier.CdwType type = new CdwPatient.CdwIdentifier.CdwType();
-      CdwPatient.CdwIdentifier.CdwType.CdwCoding coding =
-          new CdwPatient.CdwIdentifier.CdwType.CdwCoding();
-      coding.setSystem("http://test-code");
-      coding.setCode("C0D3");
-      type.getCoding().add(coding);
-      identifier.setType(type);
-      identifier.setSystem("http://test-system");
-      identifier.setValue("123456789");
-      CdwPatient.CdwIdentifier.CdwAssigner assigner = new CdwPatient.CdwIdentifier.CdwAssigner();
-      assigner.setDisplay("tester-test-index");
-      identifier.setAssigner(assigner);
-      return identifier;
-    }
-
-    CdwPatient.CdwMaritalStatus maritalStatus() {
-      CdwPatient.CdwMaritalStatus maritalStatus = new CdwPatient.CdwMaritalStatus();
-      maritalStatus.setText("testMarriage");
-      CdwPatient.CdwMaritalStatus.CdwCoding coding = new CdwPatient.CdwMaritalStatus.CdwCoding();
-      coding.setSystem(CdwMaritalStatusSystems.HTTP_HL_7_ORG_FHIR_MARITAL_STATUS);
-      coding.setCode(CdwMaritalStatusCodes.M);
-      coding.setDisplay("Married");
-      maritalStatus.getCoding().add(coding);
-      return maritalStatus;
-    }
-
-    CdwPatient.CdwName name() {
-      CdwPatient.CdwName name = new CdwPatient.CdwName();
-      name.setUse("usual");
-      name.setText("FOOMAN FOO");
-      name.setFamily("FOO");
-      name.setGiven("FOOMAN");
-      return name;
-    }
-
-    CdwExtensions raceCdwExtensions() {
-      CdwExtensions CdwExtensions = new CdwExtensions();
-      CdwExtensions.setUrl("http://test-race");
-
-      CdwExtension extension1 = new CdwExtension();
-      extension1.setUrl("ombTest");
-
-      CdwExtension.CdwValueCoding valueCoding = new CdwExtension.CdwValueCoding();
-      valueCoding.setSystem("http://test-race");
-      valueCoding.setCode("R4C3");
-      valueCoding.setDisplay("tester");
-      extension1.setValueCoding(valueCoding);
-
-      CdwExtensions.getExtension().add(extension1);
-
-      CdwExtension extension2 = new CdwExtension();
-      extension2.setUrl("text");
-      extension2.setValueString("tester");
-
-      CdwExtensions.getExtension().add(extension2);
-      return CdwExtensions;
-    }
-
-    CdwPatient.CdwTelecoms telecoms() {
-      CdwPatient.CdwTelecoms testTelecoms = new CdwPatient.CdwTelecoms();
-
-      CdwPatient.CdwTelecoms.CdwTelecom telecom1 = new CdwPatient.CdwTelecoms.CdwTelecom();
-      telecom1.setSystem(CdwContactPointSystemCodes.PHONE);
-      telecom1.setValue("9998886666");
-      telecom1.setUse(CdwContactPointUseCodes.HOME);
-      testTelecoms.getTelecom().add(telecom1);
-
-      CdwPatient.CdwTelecoms.CdwTelecom telecom2 = new CdwPatient.CdwTelecoms.CdwTelecom();
-      telecom2.setSystem(CdwContactPointSystemCodes.PHONE);
-      telecom2.setValue("1112224444");
-      telecom2.setUse(CdwContactPointUseCodes.WORK);
-      testTelecoms.getTelecom().add(telecom2);
-      return testTelecoms;
-    }
   }
 }
